put messaging into GeoConcerns namespace
